refactor on viewProject button actions
